# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Hospital Management System.",
        "title": "HMS API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ekastn",
            "email": "eka@voixes.studio"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5021",
    "basePath": "/api",
    "paths": {
...
        "/doctors": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all registered doctors.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctors"
                ],
                "summary": "Get all doctors",
                "responses": {
                    "200": {
                        "description": "List of doctors",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.DoctorDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve doctors",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new doctor entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctors"
                ],
                "summary": "Create a new doctor",
                "parameters": [
                    {
                        "description": "Doctor object to be created",
                        "name": "doctor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateDoctorRequet"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Doctor created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create doctor",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/doctors/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a single doctor by their ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctors"
                ],
                "summary": "Get doctor by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Doctor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.DoctorDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Doctor not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve doctor",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update details of an existing doctor.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctors"
                ],
                "summary": "Update an existing doctor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Doctor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Doctor object with updated fields",
                        "name": "doctor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateDoctorRequet"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Doctor updated successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Doctor not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update doctor",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a doctor entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctors"
                ],
                "summary": "Delete a doctor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Doctor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Doctor deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete doctor",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/doctors/{id}/detail": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve detailed information for a single doctor, including recent patients.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctors"
                ],
                "summary": "Get detailed doctor information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Doctor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Doctor details with recent patients",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.DoctorDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Doctor not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve doctor details",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateDoctorRequet": {
            "description": "Request body for creating a new doctor",
            "type": "object",
            "required": [
                "email",
                "name",
                "phone",
                "specialty"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "jane.smith@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Dr. Jane Smith"
                },
                "phone": {
                    "type": "string",
                    "example": "+1987654321"
                },
                "specialty": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Pediatrics"
                }
            }
        },
...
        "domain.DoctorDTO": {
            "description": "Doctor data transfer object",
            "type": "object",
            "properties": {
                "availability": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TimeSlot"
                    }
                },
                "createdAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000001"
                },
                "name": {
                    "type": "string",
                    "example": "Dr. John Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "1234567890"
                },
                "specialty": {
                    "type": "string",
                    "example": "Cardiology"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                }
            }
        },
        "domain.DoctorDetailResponse": {
            "description": "Detailed doctor information including recent patients",
            "type": "object",
            "properties": {
                "doctor": {
                    "$ref": "#/definitions/domain.DoctorDTO"
                },
                "recentPatients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.PatientDTO"
                    }
                }
            }
        },
...
        "domain.PatientDTO": {
            "description": "Patient data transfer object",
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "123 Main St"
                },
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string",
                    "example": "jane.doe@example.com"
                },
                "gender": {
                    "type": "string",
                    "example": "Female"
                },
                "id": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000001"
                },
                "lastVisit": {
                    "type": "string",
                    "example": "2025-07-17T10:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "1234567890"
                }
            }
        },
...
        "domain.TimeSlot": {
            "type": "object",
            "properties": {
                "dayOfWeek": {
                    "description": "0-6 (Sunday-Saturday)",
                    "type": "integer"
                },
                "endTime": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
...
        "domain.UpdateDoctorRequet": {
            "description": "Request body for updating an existing doctor",
            "type": "object",
            "required": [
                "email",
                "name",
                "phone",
                "specialty"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "jane.smith@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Dr. Jane Smith-Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "+1987654321"
                },
                "specialty": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Pediatrics"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Error message"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}